+++ crates/stages/src/stages/mod.rs
@@ -21,8 +21,6 @@ mod sender_recovery;
 /// The transaction lookup stage
 mod tx_lookup;
 
-use std::{collections::HashMap, ops::RangeBounds};
-
 pub use bodies::*;
 pub use execution::*;
 pub use finish::*;
@@ -32,191 +30,12 @@ pub use headers::*;
 pub use index_account_history::*;
 pub use index_storage_history::*;
 pub use merkle::*;
-use reth_db::{
-    cursor::{DbCursorRO, DbCursorRW},
-    models::sharded_key::NUM_OF_INDICES_IN_SHARD,
-    table::{Decompress, Table},
-    transaction::{DbTx, DbTxMut},
-    BlockNumberList,
-};
-use reth_etl::Collector;
-use reth_interfaces::provider::ProviderResult;
-use reth_primitives::BlockNumber;
+
 pub use sender_recovery::*;
-use std::hash::Hash;
 pub use tx_lookup::*;
 
-pub(crate) enum LoadMode {
-    KeepLast,
-    Flush,
-}
-
-impl LoadMode {
-    fn is_flush(&self) -> bool {
-        matches!(self, Self::Flush)
-    }
-}
-pub(crate) fn collect_history_indices<TX, CHANGESET, HISTORY, CACHEKEY>(
-    tx: &TX,
-    range: impl RangeBounds<CHANGESET::Key>,
-    sharded_key_factory: impl Fn(CACHEKEY, BlockNumber) -> HISTORY::Key,
-    cache_key_factory: impl Fn((CHANGESET::Key, CHANGESET::Value)) -> (u64, CACHEKEY),
-) -> ProviderResult<Collector<HISTORY::Key, HISTORY::Value>>
-where
-    TX: DbTxMut + DbTx,
-    CHANGESET: Table,
-    HISTORY: Table<Value = BlockNumberList>,
-    CACHEKEY: Eq + Hash,
-{
-    let mut changeset_cursor = tx.cursor_read::<CHANGESET>()?;
-
-    let mut collector = Collector::new(500 * 1024 * 1024, None);
-    let mut cache: HashMap<CACHEKEY, Vec<u64>> = HashMap::new();
-    let mut entries = 0;
-    let entries_threshold = 100_000;
-
-    let mut collect = |cache: HashMap<CACHEKEY, Vec<u64>>| {
-        for (key, indice_list) in cache {
-            let last = indice_list.last().expect("qed");
-            collector
-                .insert(
-                    sharded_key_factory(key, *last),
-                    BlockNumberList::new_pre_sorted(indice_list),
-                )
-                .unwrap();
-        }
-    };
-
-    for entry in changeset_cursor.walk_range(range)? {
-        let (index, key) = cache_key_factory(entry?);
-        cache.entry(key).or_default().push(index);
-
-        entries += 1;
-        if entries > entries_threshold {
-            entries = 0;
-
-            collect(cache);
-            cache = HashMap::default();
-        }
-    }
-    collect(cache);
-
-    Ok(collector)
-}
-
-/// Given a list on indices we push them to DB after splitting them into shards.
-pub(crate) fn load_indices<T, C, P>(
-    cursor: &mut C,
-    partial_key: P,
-    list: &mut Vec<BlockNumber>,
-    sharded_key_factory: &impl Fn(P, BlockNumber) -> <T as Table>::Key,
-    append_only: bool,
-    mode: LoadMode,
-) -> ProviderResult<()>
-where
-    T: Table<Value = BlockNumberList>,
-    P: Copy,
-    C: DbCursorRO<T> + DbCursorRW<T>,
-{
-    if list.len() > NUM_OF_INDICES_IN_SHARD || mode.is_flush() {
-        let chunks = list
-            .chunks(NUM_OF_INDICES_IN_SHARD)
-            .map(|chunks| chunks.to_vec())
-            .collect::<Vec<Vec<u64>>>();
-
-        let mut iter = chunks.into_iter().peekable();
-        while let Some(chunk) = iter.next() {
-            let highest = *chunk.last().unwrap();
-
-            if !mode.is_flush() && iter.peek().is_none() {
-                *list = chunk;
-            } else {
-                let key = sharded_key_factory(partial_key, highest);
-                let value = BlockNumberList::new_pre_sorted(chunk);
-
-                if append_only {
-                    cursor.append(key, value)?;
-                } else {
-                    cursor.upsert(key, value)?;
-                }
-            }
-        }
-    }
-
-    Ok(())
-}
-
-pub(crate) fn load_history_indices<TX, HISTORY, P>(
-    tx: &TX,
-    mut collector: Collector<HISTORY::Key, HISTORY::Value>,
-    append_only: bool,
-    sharded_key_factory: impl Clone + Fn(P, u64) -> <HISTORY as Table>::Key,
-    decode_key: impl Fn(Vec<u8>) -> ProviderResult<<HISTORY as Table>::Key>,
-    get_partial: impl Fn(<HISTORY as Table>::Key) -> P,
-) -> ProviderResult<()>
-where
-    TX: DbTxMut + DbTx,
-    P: Copy + Default + Eq,
-    HISTORY: Table<Value = BlockNumberList>,
-{
-    let mut write_cursor = tx.cursor_write::<HISTORY>()?;
-    let mut current_partial = P::default();
-    let mut current_list = Vec::<u64>::new();
-
-    for element in collector.iter().unwrap() {
-        let (k, v) = element.unwrap();
-        let sharded = decode_key(k).unwrap();
-        let new_list = BlockNumberList::decompress_owned(v).unwrap();
-        let partial = get_partial(sharded);
-        if current_partial != partial {
-            // We have reached the end of this subset of keys (eg. Addresses on AccountHistory), so
-            // we need to flush its indices.
-            load_indices(
-                &mut write_cursor,
-                current_partial,
-                &mut current_list,
-                &sharded_key_factory,
-                append_only,
-                LoadMode::Flush,
-            )?;
-
-            current_partial = partial;
-            current_list.clear();
-
-            // If it's not the first sync, there might an existing shard already, so we need to
-            // merge it.
-            if !append_only {
-                if let Some((_, last_database_shard)) =
-                    write_cursor.seek_exact(sharded_key_factory(current_partial, u64::MAX))?
-                {
-                    current_list.extend(last_database_shard.iter());
-                }
-            }
-        }
-
-        current_list.extend(new_list.iter());
-        load_indices(
-            &mut write_cursor,
-            current_partial,
-            &mut current_list,
-            &sharded_key_factory,
-            append_only,
-            LoadMode::KeepLast,
-        )?;
-    }
-
-    // There will be one remaining list of indices that needs to be flushed to DB.
-    load_indices(
-        &mut write_cursor,
-        current_partial,
-        &mut current_list,
-        &sharded_key_factory,
-        append_only,
-        LoadMode::Flush,
-    )?;
-
-    Ok(())
-}
+mod utils;
+use utils::*;
 
 #[cfg(test)]
 mod tests {
