+++ crates/stages/src/stages/utils.rs
@@ -1,0 +1,216 @@
+//! Utils for `stages`.
+use crate::StageError;
+use reth_db::{
+    cursor::{DbCursorRO, DbCursorRW},
+    models::sharded_key::NUM_OF_INDICES_IN_SHARD,
+    table::{Decompress, Table},
+    transaction::{DbTx, DbTxMut},
+    BlockNumberList, DatabaseError,
+};
+use reth_etl::Collector;
+use reth_primitives::BlockNumber;
+use std::{collections::HashMap, hash::Hash, ops::RangeBounds};
+
+/// Collects all history (`H`) indices for a range of changesets (`CS`) and stores them in a
+/// [`Collector`].
+///
+/// ## Process
+/// The function utilizes a `HashMap` cache with a structure of `PartialKey` (`P`) (Address or
+/// Address.StorageKey) to `BlockNumberList`. When the cache exceeds its capacity, its contents are
+/// moved to a [`Collector`]. Here, each entry's key is a concatenation of `PartialKey` and the
+/// highest block number in its list.
+///
+/// ## Example
+/// 1. Initial Cache State: `{ Address1: [1,2,3], ... }`
+/// 2. Cache is flushed to the `Collector`.
+/// 3. Updated Cache State: `{ Address1: [100,300], ... }`
+/// 4. Cache is flushed again.
+///
+/// As a result, the `Collector` will contain entries such as `(Address1.3, [1,2,3])` and
+/// `(Address1.300, [100,300])`. The entries may be stored across one or more files.
+pub(crate) fn collect_history_indices<TX, CS, H, P>(
+    tx: &TX,
+    range: impl RangeBounds<CS::Key>,
+    sharded_key_factory: impl Fn(P, BlockNumber) -> H::Key,
+    partial_key_factory: impl Fn((CS::Key, CS::Value)) -> (u64, P),
+) -> Result<Collector<H::Key, H::Value>, StageError>
+where
+    TX: DbTxMut + DbTx,
+    CS: Table,
+    H: Table<Value = BlockNumberList>,
+    P: Eq + Hash,
+{
+    let mut changeset_cursor = tx.cursor_read::<CS>()?;
+
+    let mut collector = Collector::new(500 * 1024 * 1024, None);
+    let mut cache: HashMap<P, Vec<u64>> = HashMap::new();
+    let mut entries = 0;
+    let entries_threshold = 100_000;
+
+    let mut collect = |cache: HashMap<P, Vec<u64>>| {
+        for (key, indice_list) in cache {
+            let last = indice_list.last().expect("qed");
+            collector.insert(
+                sharded_key_factory(key, *last),
+                BlockNumberList::new_pre_sorted(indice_list),
+            )?;
+        }
+        Ok::<(), StageError>(())
+    };
+
+    for entry in changeset_cursor.walk_range(range)? {
+        let (index, key) = partial_key_factory(entry?);
+        cache.entry(key).or_default().push(index);
+
+        entries += 1;
+        if entries > entries_threshold {
+            entries = 0;
+
+            collect(cache)?;
+            cache = HashMap::default();
+        }
+    }
+    collect(cache)?;
+
+    Ok(collector)
+}
+
+/// Given a [`Collector`] created by [`collect_history_indices`] it iterates all entries, loading
+/// the indices into the database in shards.
+///
+///  ## Process
+/// Iterates over elements, grouping indices by their partial keys (e.g., `Address` or
+/// `Address.StorageKey`). It flushes indices to disk when reaching a shard's max length
+/// (`NUM_OF_INDICES_IN_SHARD`) or when the partial key changes, ensuring the last previous partial
+/// key shard is stored.
+pub(crate) fn load_history_indices<TX, H, P>(
+    tx: &TX,
+    mut collector: Collector<H::Key, H::Value>,
+    append_only: bool,
+    sharded_key_factory: impl Clone + Fn(P, u64) -> <H as Table>::Key,
+    decode_key: impl Fn(Vec<u8>) -> Result<<H as Table>::Key, DatabaseError>,
+    get_partial: impl Fn(<H as Table>::Key) -> P,
+) -> Result<(), StageError>
+where
+    TX: DbTxMut + DbTx,
+    H: Table<Value = BlockNumberList>,
+    P: Copy + Default + Eq,
+{
+    let mut write_cursor = tx.cursor_write::<H>()?;
+    let mut current_partial = P::default();
+    let mut current_list = Vec::<u64>::new();
+
+    for element in collector.iter()? {
+        let (k, v) = element?;
+        let sharded_key = decode_key(k)?;
+        let new_list = BlockNumberList::decompress_owned(v)?;
+
+        // AccountsHistory: `Address`.
+        // StorageHistory: `Address.StorageKey`.
+        let partial_key = get_partial(sharded_key);
+
+        if current_partial != partial_key {
+            // We have reached the end of this subset of keys so
+            // we need to flush its last indice shard.
+            load_indices(
+                &mut write_cursor,
+                current_partial,
+                &mut current_list,
+                &sharded_key_factory,
+                append_only,
+                LoadMode::Flush,
+            )?;
+
+            current_partial = partial_key;
+            current_list.clear();
+
+            // If it's not the first sync, there might an existing shard already, so we need to
+            // merge it with the one coming from the collector
+            if !append_only {
+                if let Some((_, last_database_shard)) =
+                    write_cursor.seek_exact(sharded_key_factory(current_partial, u64::MAX))?
+                {
+                    current_list.extend(last_database_shard.iter());
+                }
+            }
+        }
+
+        current_list.extend(new_list.iter());
+        load_indices(
+            &mut write_cursor,
+            current_partial,
+            &mut current_list,
+            &sharded_key_factory,
+            append_only,
+            LoadMode::KeepLast,
+        )?;
+    }
+
+    // There will be one remaining shard that needs to be flushed to DB.
+    load_indices(
+        &mut write_cursor,
+        current_partial,
+        &mut current_list,
+        &sharded_key_factory,
+        append_only,
+        LoadMode::Flush,
+    )?;
+
+    Ok(())
+}
+
+/// Shard and insert the indice list according to [`LoadMode`] and its length.
+pub(crate) fn load_indices<H, C, P>(
+    cursor: &mut C,
+    partial_key: P,
+    list: &mut Vec<BlockNumber>,
+    sharded_key_factory: &impl Fn(P, BlockNumber) -> <H as Table>::Key,
+    append_only: bool,
+    mode: LoadMode,
+) -> Result<(), StageError>
+where
+    C: DbCursorRO<H> + DbCursorRW<H>,
+    H: Table<Value = BlockNumberList>,
+    P: Copy,
+{
+    if list.len() > NUM_OF_INDICES_IN_SHARD || mode.is_flush() {
+        let chunks = list
+            .chunks(NUM_OF_INDICES_IN_SHARD)
+            .map(|chunks| chunks.to_vec())
+            .collect::<Vec<Vec<u64>>>();
+
+        let mut iter = chunks.into_iter().peekable();
+        while let Some(chunk) = iter.next() {
+            let highest = *chunk.last().expect("only exist when there's an indice");
+
+            if !mode.is_flush() && iter.peek().is_none() {
+                *list = chunk;
+            } else {
+                let key = sharded_key_factory(partial_key, highest);
+                let value = BlockNumberList::new_pre_sorted(chunk);
+
+                if append_only {
+                    cursor.append(key, value)?;
+                } else {
+                    cursor.upsert(key, value)?;
+                }
+            }
+        }
+    }
+
+    Ok(())
+}
+
+/// Mode on how to load indice shards into memory.
+pub(crate) enum LoadMode {
+    /// Keep the last shard in memory and don't flush it to the database
+    KeepLast,
+    /// Flush all shards into the database.
+    Flush,
+}
+
+impl LoadMode {
+    fn is_flush(&self) -> bool {
+        matches!(self, Self::Flush)
+    }
+}
